requirements.txt
fastapi
uvicorn[standard]
requests
scikit-learn
nltk
python-dotenv
